+++ modules/webhook/lambdas/webhook/src/lambda.ts
@@ -1,13 +1,18 @@
 import { handle } from './webhook/handler';
 import { APIGatewayEvent, Context, Callback } from 'aws-lambda';
+import { logger } from './webhook/logger';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export interface Response {
+  statusCode: number;
+  body?: string;
+}
+
 export const githubWebhook = async (event: APIGatewayEvent, context: Context, callback: Callback): Promise<void> => {
+  logger.setSettings({ requestId: context.awsRequestId });
+  logger.debug(JSON.stringify(event));
   try {
-    const statusCode = await handle(event.headers, event.body as string);
-    callback(null, {
-      statusCode: statusCode,
-    });
+    const response = await handle(event.headers, event.body as string);
+    callback(null, response);
   } catch (e) {
     callback(e as Error);
   }