+++ modules/webhook/lambdas/webhook/src/webhook/handler.ts
@@ -3,8 +3,13 @@ import { Webhooks } from '@octokit/webhooks';
 import { sendActionRequest } from '../sqs';
 import { CheckRunEvent, WorkflowJobEvent } from '@octokit/webhooks-types';
 import { getParameterValue } from '../ssm';
+import { logger as rootLogger, LogFields } from './logger';
+import { Response } from '../lambda';
 
-export const handle = async (headers: IncomingHttpHeaders, body: string): Promise<number> => {
+const supportedEvents = ['check_run', 'workflow_job'];
+const logger = rootLogger.getChildLogger();
+
+export async function handle(headers: IncomingHttpHeaders, body: string): Promise<Response> {
   // ensure header keys lower case since github headers can contain capitals.
   for (const key in headers) {
     headers[key.toLowerCase()] = headers[key];
@@ -12,32 +17,61 @@ export const handle = async (headers: IncomingHttpHeaders, body: string): Promis
 
   const githubEvent = headers['x-github-event'] as string;
 
-  let status = await verifySignature(githubEvent, headers['x-hub-signature'] as string, body);
-  if (status != 200) {
-    return status;
+  let response: Response = {
+    statusCode: await verifySignature(githubEvent, headers['x-hub-signature'] as string, body),
+  };
+
+  if (response.statusCode != 200) {
+    return response;
   }
   const payload = JSON.parse(body);
-  console.info(`Received Github event ${githubEvent} from ${payload.repository.full_name}`);
+  LogFields.fields.event = githubEvent;
+  LogFields.fields.repository = payload.repository.full_name;
+  LogFields.fields.action = payload.action;
+
+  if (!supportedEvents.includes(githubEvent)) {
+    logger.warn(`Unsupported event type.`, LogFields.print());
+    return {
+      statusCode: 202,
+      body: `Ignoring unsupported event ${githubEvent}`,
+    };
+  }
+
+  LogFields.fields.name = payload[githubEvent].name;
+  LogFields.fields.status = payload[githubEvent].status;
+  LogFields.fields.started_at = payload[githubEvent]?.started_at;
+
+  /*
+  The app subscribes to all `check_run` and `workflow_job` events.
+  If the event status is `completed`, log the data for workflow metrics.
+  */
+  LogFields.fields.completed_at = payload[githubEvent]?.completed_at;
+  LogFields.fields.conclusion = payload[githubEvent]?.conclusion;
 
   if (isRepoNotAllowed(payload.repository.full_name)) {
-    console.error(`Received event from unauthorized repository ${payload.repository.full_name}`);
-    return 403;
+    logger.error(`Received event from unauthorized repository ${payload.repository.full_name}`, LogFields.print());
+    return {
+      statusCode: 403,
+    };
   }
+
+  logger.info(`Processing Github event`, LogFields.print());
 
   if (githubEvent == 'workflow_job') {
-    status = await handleWorkflowJob(payload as WorkflowJobEvent, githubEvent);
+    response = await handleWorkflowJob(payload as WorkflowJobEvent, githubEvent);
   } else if (githubEvent == 'check_run') {
-    status = await handleCheckRun(payload as CheckRunEvent, githubEvent);
-  } else {
-    console.warn(`Ignoring unsupported event ${githubEvent}`);
+    response = await handleCheckRun(payload as CheckRunEvent, githubEvent);
   }
 
-  return status;
-};
+  return response;
+}
 
 async function verifySignature(githubEvent: string, signature: string, body: string): Promise<number> {
   if (!signature) {
-    console.error("Github event doesn't have signature. This webhook requires a secret to be configured.");
+    logger.error(
+      "Github event doesn't have signature. This webhook requires a secret to be configured.",
+      LogFields.print(),
+    );
     return 500;
   }
 
@@ -47,18 +81,24 @@ async function verifySignature(githubEvent: string, signature: string, body: str
     secret: secret,
   });
   if (!(await webhooks.verify(body, signature))) {
-    console.error('Unable to verify signature!');
+    logger.error('Unable to verify signature!', LogFields.print());
     return 401;
   }
   return 200;
 }
 
-async function handleWorkflowJob(body: WorkflowJobEvent, githubEvent: string): Promise<number> {
+async function handleWorkflowJob(body: WorkflowJobEvent, githubEvent: string): Promise<Response> {
   const disableCheckWorkflowJobLabelsEnv = process.env.DISABLE_CHECK_WORKFLOW_JOB_LABELS || 'false';
   const disableCheckWorkflowJobLabels = JSON.parse(disableCheckWorkflowJobLabelsEnv) as boolean;
   if (!disableCheckWorkflowJobLabels && !canRunJob(body)) {
-    console.error(`Received event contains runner labels '${body.workflow_job.labels}' that are not accepted.`);
-    return 403;
+    logger.warn(
+      `Received event contains runner labels '${body.workflow_job.labels}' that are not accepted.`,
+      LogFields.print(),
+    );
+    return {
+      statusCode: 202,
+      body: `Received event contains runner labels '${body.workflow_job.labels}' that are not accepted.`,
+    };
   }
 
   let installationId = body.installation?.id;
@@ -74,11 +114,11 @@ async function handleWorkflowJob(body: WorkflowJobEvent, githubEvent: string): P
       installationId: installationId,
     });
   }
-  console.info(`Successfully queued job for ${body.repository.full_name}`);
-  return 200;
+  logger.info(`Successfully queued job for ${body.repository.full_name}`, LogFields.print());
+  return { statusCode: 201 };
 }
 
-async function handleCheckRun(body: CheckRunEvent, githubEvent: string): Promise<number> {
+async function handleCheckRun(body: CheckRunEvent, githubEvent: string): Promise<Response> {
   let installationId = body.installation?.id;
   if (installationId == null) {
     installationId = 0;
@@ -92,8 +132,8 @@ async function handleCheckRun(body: CheckRunEvent, githubEvent: string): Promise
       installationId: installationId,
     });
   }
-  console.info(`Successfully queued job for ${body.repository.full_name}`);
-  return 200;
+  logger.info(`Successfully queued job for ${body.repository.full_name}`, LogFields.print());
+  return { statusCode: 201 };
 }
 
 function isRepoNotAllowed(repo_full_name: string): boolean {
@@ -119,10 +159,11 @@ function canRunJob(job: WorkflowJobEvent): boolean {
 
   const runnerMatch = customWorkflowJobLabels.every((l) => runnerLabels.has(l));
 
-  console.debug(
+  logger.debug(
     `Received workflow job event with labels: '${JSON.stringify(job.workflow_job.labels)}'. The event does ${
       runnerMatch ? '' : 'NOT '
     }match the configured labels: '${Array.from(runnerLabels).join(',')}'`,
+    LogFields.print(),
   );
   return runnerMatch;
 }